Trekk ut dataklasser som blir brukt til å generere vedtaksperioder til egen fil
