+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/vedtaksperiode/produsent/GrunnlagForGjeldendeOgForrigeBehandling.kt
@@ -1,0 +1,6 @@
+package no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.produsent
+
+data class GrunnlagForGjeldendeOgForrigeBehandling(
+    val gjeldende: GrunnlagForPerson?,
+    val forrige: GrunnlagForPerson?
+)
