+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/vedtaksperiode/produsent/GrunnlagForVedtaksperioder.kt
@@ -1,0 +1,17 @@
+package no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.produsent
+
+import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
+import no.nav.familie.ba.sak.kjerne.endretutbetaling.domene.EndretUtbetalingAndel
+import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.Kompetanse
+import no.nav.familie.ba.sak.kjerne.fagsak.FagsakType
+import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.PersonResultat
+
+data class GrunnlagForVedtaksperioder(
+    val persongrunnlag: PersonopplysningGrunnlag,
+    val personResultater: Set<PersonResultat>,
+    val fagsakType: FagsakType,
+    val kompetanser: List<Kompetanse>,
+    val endredeUtbetalinger: List<EndretUtbetalingAndel>,
+    val andelerTilkjentYtelse: List<AndelTilkjentYtelse>
+)
